Move JupyterLab and JupyterClient into their own files.
